+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## Version 0.0.3 (2023-09-25)
+
+### Changes
+
+- Synapsis.ConcreteTypes now handles EntityBundle.
+- Synapsis.ConcreteTypes.get() will always return Synapsis.ConcreteTypes.UNKNOWN if the ConcreteType cannot be found.
+
 ## Version 0.0.2 (2023-09-22)
 
 ### Changes